requirements.txt
# Core Dependencies
numpy>=1.21.0
scipy>=1.7.0
sympy>=1.9
pandas>=1.3.0

# Speech Recognition (Optional - for voice features)
# Note: Install manually if voice features needed:
# pip install SpeechRecognition pyaudio pyttsx3

# Database and Storage
pyyaml>=6.0
python-dateutil>=2.8.0
schedule>=1.2.0

# System Integration
psutil>=5.9.0

# Scientific Computing (Optional)
# astropy>=5.0.0
# skyfield>=1.40

# GUI Framework (Built-in with Python)
# tkinter is included with Python
# customtkinter>=5.0.0  # Optional for modern UI

# Utilities
pathlib  # Built-in with Python 3.4+
sqlite3  # Built-in with Python
json  # Built-in with Python
logging  # Built-in with Python
threading  # Built-in with Python
multiprocessing  # Built-in with Python
